Minor feature changes
Disabled !remove when a pug is in session/picking mode.
Update names on sidelist on nick change.
Removed random key for Welcome in $data to look for the end of the MOTD
on the server.
Added usleep($delay_send) - $delay_send var should be added to
config.php
Added a timeout for people who backout of a started session (DISABLED
ATM)
Made !remove work with the sidelist.