maxvote
I wrote a bit of code the handle voting to change the max team size.  I
didn't have PHP on my laptop while I was writing this so I haven't been
able to test any of this yet, but it shouldn't have too many errors.  I
also need help wiring it up to the !maxvote command.